run_demo_*.py regrouped in single run.py
In order to limit the number of files the run_demo_*.py runtime
helpers are now regrouped in PythonCallingDocker/run.py.
