+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/DemoStatic/run_lyon_metropole_dowload_and_sanitize_static.py
@@ -3,8 +3,7 @@ import sys
 import demo_workflow_static as workflow
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from run_demo_lyon_metropole_dowload_and_sanitize \
-    import run_demo_lyon_metropole_dowload_and_sanitize
+from run import run_demo_lyon_metropole_dowload_and_sanitize
 
 
 if __name__ == '__main__':
