+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/DemoStatic/run_strip_attributes_static.py
@@ -3,7 +3,7 @@ import sys
 import demo_workflow_static as workflow
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from run_demo_strip_attributes import run_demo_strip_attributes
+from run import run_demo_strip_attributes
 
 
 if __name__ == '__main__':
