+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/DemoTemporal/run_split_buildings_temporal.py
@@ -3,7 +3,7 @@ import sys
 import demo_workflow_temporal as workflow
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from run_demo_split_buildings import run_demo_split_buildings
+from run import run_demo_split_buildings
 
 
 if __name__ == '__main__':
